Modularized all CRUD for room-moderator.js
