+++ index.js
@@ -15,6 +15,7 @@ const adminRouter = require('./routes/admin');
 const roomRouter = require('./routes/room-route');
 const modRouter = require('./routes/flag-route');
 const searchRouter = require('./routes/search-route');
+const moderatorRouter = require('./routes/room-moderator');
 
 const User = require('./models/user');
 
@@ -24,18 +25,18 @@ const FRONTEND_URL =
   process.env.FRONTEND_DEPLOYED_URL || 'http://localhost:3000';
 const PORT = process.env.PORT || 5000;
 app.all('/*', function(req, res, next) {
-    res.header("Access-Control-Allow-Origin", "*");
-    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
-    next();
+  res.header('Access-Control-Allow-Origin', '*');
+  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
+  next();
 });
 app.use(express.json());
 app.use(helmet());
 app.use((req, res, next) => {
   res.header('Access-Control-Allow-Credentials', 'true');
   res.header('Access-Control-Allow-Origin', 'https://main.d37zm5ayhfot8q.amplifyapp.com');
-  res.header("Access-Control-Allow-Origin", "*");
-  res.header("Access-Control-Allow-Methods", "GET,HEAD,OPTIONS,POST,PUT");
-  res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, Authorization");
+  res.header('Access-Control-Allow-Origin', '*');
+  res.header('Access-Control-Allow-Methods', 'GET,HEAD,OPTIONS,POST,PUT');
+  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Authorization');
   next();
 });
 app.use(
@@ -54,7 +55,7 @@ app.use(
     cookie: {
       httpOnly: true,
       maxAge: 1000 * 60 * 60 * 24 * 7,
-      secure: false, // Set to true once in production
+      secure: false, // Set to true once in production // This is causing the auth fail when true (only used over https?)
       SameSite: 'none',
     },
     store: new knexSessionStore({
@@ -88,6 +89,7 @@ app.use('/api/room', verifyRole, roomRouter);
 app.use('/api/admin', verifyRole, adminRouter);
 app.use('/api/mod', verifyRole, modRouter);
 app.use('/api/search', verifyRole, searchRouter);
+app.use('/api/moderator', verifyRole, moderatorRouter);
 
 app.get('/', (request, response) =>
   response.send({ message: 'Server working' })
